Fixed if/then logic and check_output
Before, even if there was verification failure (couldn't even verify—not that the verification had a mismatch), the script would still prompt to trust changes. Now it will skip all that if there's a verification failure. 

Also added in an `stderr=subprocess.STDOUT` parameter to `check_output`, because it was failing on the AutoPkg 1.1 release candidate.